refactor(bench): split flat cmd/bench into internal/dfsbench packages
Move the harness from one package main into small cohesive packages
(exec, config, fio, report, backend, cloud, run) under internal/dfsbench,
each owning its own cobra command; cmd/bench/main.go just wires them.
Every file is now <=211 lines. Pure move/split, no behavior change:
gofmt/vet/golangci-lint clean, all 23 tests pass, CLI intact.
