+++ internal/dfsbench/cloud/cloud_test.go
@@ -1,4 +1,4 @@
-package main
+package cloud
 
 import (
 	"os"
@@ -24,7 +24,7 @@ func TestVMState_RoundTrip(t *testing.T) {
 	if err := saveVM(want); err != nil {
 		t.Fatal(err)
 	}
-	got, err := loadVM()
+	got, err := LoadVM()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -42,43 +42,8 @@ func TestVMState_RoundTrip(t *testing.T) {
 	}
 }
 
-func TestRemoteRunArgs_ForwardsSelection(t *testing.T) {
-	f := &runFlags{
-		systems:    []string{"dittofs-s3", "s3fs"},
-		workloads:  []string{"seq-read"},
-		sizes:      []string{"large"},
-		evictCache: true,
-		resume:     true,
-	}
-	got := remoteRunArgs(f)
-	for _, want := range []string{
-		"/root/dfsbench run",
-		"--config /root/dfsbench.yaml",
-		"--results /root/bench-results",
-		"--systems dittofs-s3,s3fs",
-		"--workloads seq-read",
-		"--sizes large",
-		"--resume",
-	} {
-		if !strings.Contains(got, want) {
-			t.Errorf("remote args missing %q in: %s", want, got)
-		}
-	}
-	// evictCache true is the default → no explicit flag.
-	if strings.Contains(got, "--evict-cache") {
-		t.Errorf("evict-cache default should not be forwarded: %s", got)
-	}
-}
-
-func TestRemoteRunArgs_EvictDisabled(t *testing.T) {
-	got := remoteRunArgs(&runFlags{systems: []string{"local-disk"}})
-	if !strings.Contains(got, "--evict-cache=false") {
-		t.Errorf("disabled evict must be forwarded explicitly: %s", got)
-	}
-}
-
 func TestBuildDriver_DetachContract(t *testing.T) {
-	d := buildDriver("/root/dfsbench run --systems local-disk")
+	d := BuildDriver("/root/dfsbench run --systems local-disk")
 	// The three properties that make a dropped ssh session survivable:
 	for _, want := range []string{
 		". /root/bench.env",    // creds sourced, not on argv
