+++ internal/dfsbench/exec/ssh_test.go
@@ -1,4 +1,4 @@
-package main
+package exec
 
 import (
 	"reflect"
