+++ internal/dfsbench/fio/workload.go
@@ -1,4 +1,4 @@
-package main
+package fio
 
 import (
 	"fmt"
@@ -7,10 +7,10 @@ import (
 	"github.com/marmos91/dittofs/bench/workloads"
 )
 
-// knownWorkloads are the fio job files shipped in bench/workloads. Each maps to
+// KnownWorkloads are the fio job files shipped in bench/workloads. Each maps to
 // <name>.fio. Kept as an explicit list (not a dir scan) so --workloads
 // selection and `list` have a stable, ordered set.
-var knownWorkloads = []string{
+var KnownWorkloads = []string{
 	"seq-write",
 	"seq-read",
 	"rand-write-4k",
@@ -19,20 +19,21 @@ var knownWorkloads = []string{
 	"metadata",
 }
 
-// sizeClasses map a friendly size name to the fio --size value. Explicit sizes
+// SizeClasses map a friendly size name to the fio --size value. Explicit sizes
 // (e.g. "512m") are also accepted and passed through verbatim.
-var sizeClasses = map[string]string{
+var SizeClasses = map[string]string{
 	"small":  "64k",
 	"medium": "1m",
 	"large":  "1g",
 }
 
-func sizeClassOrder() []string { return []string{"small", "medium", "large"} }
+// SizeClassOrder is the display order for the named size classes.
+func SizeClassOrder() []string { return []string{"small", "medium", "large"} }
 
-// resolveSize returns the fio --size string for a size selector, accepting
+// ResolveSize returns the fio --size string for a size selector, accepting
 // either a named class or an explicit fio size literal.
-func resolveSize(sel string) string {
-	if v, ok := sizeClasses[sel]; ok {
+func ResolveSize(sel string) string {
+	if v, ok := SizeClasses[sel]; ok {
 		return v
 	}
 	return sel
@@ -59,10 +60,10 @@ func jobDefaults(engine, dir, size string, threads, runtime int) map[string]stri
 
 var varRe = regexp.MustCompile(`\$\{([A-Za-z_][A-Za-z0-9_]*)(?::-([^}]*))?\}`)
 
-// expandJob renders a .fio template, replacing ${VAR} and ${VAR:-default} with
+// ExpandJob renders a .fio template, replacing ${VAR} and ${VAR:-default} with
 // values from params, falling back to the template's own default. Unset vars
 // with no default expand to empty (matching shell).
-func expandJob(tmpl string, params map[string]string) string {
+func ExpandJob(tmpl string, params map[string]string) string {
 	return varRe.ReplaceAllStringFunc(tmpl, func(m string) string {
 		g := varRe.FindStringSubmatch(m)
 		name, def := g[1], g[2]
@@ -79,12 +80,12 @@ func loadJob(name string, params map[string]string) (string, error) {
 	if err != nil {
 		return "", fmt.Errorf("unknown workload %q: %w", name, err)
 	}
-	return expandJob(string(raw), params), nil
+	return ExpandJob(string(raw), params), nil
 }
 
-// validWorkload reports whether name is a shipped workload.
-func validWorkload(name string) bool {
-	for _, w := range knownWorkloads {
+// ValidWorkload reports whether name is a shipped workload.
+func ValidWorkload(name string) bool {
+	for _, w := range KnownWorkloads {
 		if w == name {
 			return true
 		}
